Deduplicate code for listening to Item changes
Also add error handling.

Closes #3654
